Remove wrong comments and initialize multi attn.
Some comments in the file indicate an error which is derived from
multiplying with the word itself in the query, key product.
That's actually not an error because the word should pay attention to
itself; consider a case in which all the other words in context are
irrelevant.
Also, multiheaded attention is initialized